# gen/openapiv2/services.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "weewar/v1/models.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/maps": {
      "get": {
        "summary": "ListMaps returns all available maps",
        "operationId": "MapsService_ListMaps",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListMapsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pagination.pageKey",
            "description": "*\nInstead of an offset an abstract  \"page\" key is provided that offers\nan opaque \"pointer\" into some offset in a result set.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pagination.pageOffset",
            "description": "*\nIf a pagekey is not supported we can also support a direct integer offset\nfor cases where it makes sense.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pagination.pageSize",
            "description": "*\nNumber of results to return.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "ownerId",
            "description": "May be filter by owner id",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "MapsService"
        ]
      },
      "post": {
        "summary": "*\nCreate a new map",
        "operationId": "MapsService_CreateMap",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateMapResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateMapRequest"
            }
          }
        ],
        "tags": [
          "MapsService"
        ]
      }
    },
    "/v1/maps/{id}": {
      "get": {
        "summary": "GetMap returns a specific map with metadata",
        "operationId": "MapsService_GetMap",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetMapResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "version",
            "description": "Optional, defaults to default_version",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "MapsService"
        ]
      },
      "delete": {
        "summary": "*\nDelete a particular map",
        "operationId": "MapsService_DeleteMap",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteMapResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "*\nID of the map to be deleted.",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": "[^/]+"
          }
        ],
        "tags": [
          "MapsService"
        ]
      }
    },
    "/v1/maps/{map.id}": {
      "patch": {
        "summary": "GetMap returns a specific map with metadata",
        "operationId": "MapsService_UpdateMap",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateMapResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "map.id",
            "description": "Unique ID for the map",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": "[^/]+"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "map": {
                  "type": "object",
                  "properties": {
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "name": {
                      "type": "string",
                      "title": "Name if items have names"
                    },
                    "description": {
                      "type": "string",
                      "title": "Description if map has a description"
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "title": "Some tags"
                    },
                    "imageUrl": {
                      "type": "string",
                      "title": "A possible image url"
                    },
                    "difficulty": {
                      "type": "string",
                      "title": "Difficulty - example attribute"
                    }
                  },
                  "title": "*\nMap being updated"
                },
                "updateMask": {
                  "type": "string",
                  "description": "*\nMask of fields being updated in this Map to make partial changes."
                }
              },
              "title": "UpdateMapRequest"
            }
          }
        ],
        "tags": [
          "MapsService"
        ]
      }
    },
    "/v1/maps:batchGet": {
      "get": {
        "summary": "*\nBatch get multiple maps by ID",
        "operationId": "MapsService_GetMaps",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetMapsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ids",
            "description": "*\nIDs of the map to be fetched",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "tags": [
          "MapsService"
        ]
      }
    },
    "/v1/users": {
      "get": {
        "summary": "ListUsers returns all available users",
        "operationId": "UsersService_ListUsers",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListUsersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pagination.pageKey",
            "description": "*\nInstead of an offset an abstract  \"page\" key is provided that offers\nan opaque \"pointer\" into some offset in a result set.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pagination.pageOffset",
            "description": "*\nIf a pagekey is not supported we can also support a direct integer offset\nfor cases where it makes sense.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pagination.pageSize",
            "description": "*\nNumber of results to return.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "ownerId",
            "description": "May be filter by owner id",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "UsersService"
        ]
      },
      "post": {
        "summary": "*\nCreate a new user",
        "operationId": "UsersService_CreateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateUserRequest"
            }
          }
        ],
        "tags": [
          "UsersService"
        ]
      }
    },
    "/v1/users/{id}": {
      "get": {
        "summary": "GetUser returns a specific user with metadata",
        "operationId": "UsersService_GetUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "version",
            "description": "Optional, defaults to default_version",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "UsersService"
        ]
      },
      "delete": {
        "summary": "*\nDelete a particular user",
        "operationId": "UsersService_DeleteUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "*\nID of the user to be deleted.",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": "[^/]+"
          }
        ],
        "tags": [
          "UsersService"
        ]
      }
    },
    "/v1/users/{user.id}": {
      "patch": {
        "summary": "GetUser returns a specific user with metadata",
        "operationId": "UsersService_UpdateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "user.id",
            "description": "Unique ID for the user",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": "[^/]+"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "object",
                  "properties": {
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "name": {
                      "type": "string",
                      "title": "Name if items have names"
                    },
                    "description": {
                      "type": "string",
                      "title": "Description if user has a description"
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "title": "Some tags"
                    },
                    "imageUrl": {
                      "type": "string",
                      "title": "A possible image url"
                    },
                    "difficulty": {
                      "type": "string",
                      "title": "Difficulty - example attribute"
                    }
                  },
                  "title": "*\nUser being updated"
                },
                "updateMask": {
                  "type": "string",
                  "description": "*\nMask of fields being updated in this User to make partial changes."
                }
              },
              "title": "UpdateUserRequest"
            }
          }
        ],
        "tags": [
          "UsersService"
        ]
      }
    },
    "/v1/users:batchGet": {
      "get": {
        "summary": "*\nBatch get multiple users by ID",
        "operationId": "UsersService_GetUsers",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetUsersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ids",
            "description": "*\nIDs of the user to be fetched",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "tags": [
          "UsersService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1CreateMapRequest": {
      "type": "object",
      "properties": {
        "map": {
          "$ref": "#/definitions/v1Map",
          "title": "*\nMap being updated"
        }
      },
      "title": "*\nMap creation request object"
    },
    "v1CreateMapResponse": {
      "type": "object",
      "properties": {
        "map": {
          "$ref": "#/definitions/v1Map",
          "title": "*\nMap being created"
        },
        "fieldErrors": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "description": "*\nError specific to a field if there are any errors."
        }
      },
      "description": "*\nResponse of an map creation."
    },
    "v1CreateUserRequest": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/v1User",
          "title": "*\nUser being updated"
        }
      },
      "title": "*\nUser creation request object"
    },
    "v1CreateUserResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/v1User",
          "title": "*\nUser being created"
        },
        "fieldErrors": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "description": "*\nError specific to a field if there are any errors."
        }
      },
      "description": "*\nResponse of an user creation."
    },
...
    "v1DeleteMapResponse": {
      "type": "object",
      "title": "*\nMap deletion response"
    },
    "v1DeleteUserResponse": {
      "type": "object",
      "title": "*\nUser deletion response"
    },
...
    "v1GetMapResponse": {
      "type": "object",
      "properties": {
        "map": {
          "$ref": "#/definitions/v1Map"
        }
      }
    },
    "v1GetMapsResponse": {
      "type": "object",
      "properties": {
        "maps": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/v1Map"
          }
        }
      },
      "title": "*\nMap batch-get response"
    },
    "v1GetUserResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/v1User"
        }
      }
    },
    "v1GetUsersResponse": {
      "type": "object",
      "properties": {
        "users": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/v1User"
          }
        }
      },
      "title": "*\nUser batch-get response"
    },
...
    "v1ListMapsResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Map"
          }
        },
        "pagination": {
          "$ref": "#/definitions/v1PaginationResponse"
        }
      }
    },
    "v1ListUsersResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1User"
          }
        },
        "pagination": {
          "$ref": "#/definitions/v1PaginationResponse"
        }
      }
    },
    "v1Map": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "string",
          "title": "Unique ID for the map"
        },
        "name": {
          "type": "string",
          "title": "Name if items have names"
        },
        "description": {
          "type": "string",
          "title": "Description if map has a description"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Some tags"
        },
        "imageUrl": {
          "type": "string",
          "title": "A possible image url"
        },
        "difficulty": {
          "type": "string",
          "title": "Difficulty - example attribute"
        }
      }
    },
...
    "v1PaginationResponse": {
      "type": "object",
      "properties": {
        "nextPageKey": {
          "type": "string",
          "description": "*\nThe key/pointer string that subsequent List requests should pass to\ncontinue the pagination."
        },
        "nextPageOffset": {
          "type": "integer",
          "format": "int32",
          "title": "*\nAlso support an integer offset if possible"
        },
        "hasMore": {
          "type": "boolean",
          "description": "*\nWhether theere are more results."
        },
        "totalResults": {
          "type": "integer",
          "format": "int32",
          "description": "*\nTotal number of results."
        }
      }
    },
...
    "v1UpdateMapResponse": {
      "type": "object",
      "properties": {
        "map": {
          "$ref": "#/definitions/v1Map",
          "title": "*\nMap being updated"
        }
      },
      "description": "*\nThe request for (partially) updating an Map.",
      "title": "UpdateMapResponse"
    },
    "v1UpdateUserResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/v1User",
          "title": "*\nUser being updated"
        }
      },
      "description": "*\nThe request for (partially) updating an User.",
      "title": "UpdateUserResponse"
    },
    "v1User": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "string",
          "title": "Unique ID for the user"
        },
        "name": {
          "type": "string",
          "title": "Name if items have names"
        },
        "description": {
          "type": "string",
          "title": "Description if user has a description"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Some tags"
        },
        "imageUrl": {
          "type": "string",
          "title": "A possible image url"
        },
        "difficulty": {
          "type": "string",
          "title": "Difficulty - example attribute"
        }
      }
    }
  }
}